🧪 Update simple-cpi-6 and workflow
