+++ examples/cpi/simple-cpi-6/Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["programs/*", "trident-tests/fuzz_tests"]
+members = ["programs/*"]
 resolver = "2"
 
 [profile.release]
