+++ examples/cpi/simple-cpi-6/programs/callee/src/lib.rs
@@ -6,11 +6,11 @@ declare_id!("HJR1TK8bgrUWzysdpS1pBGBYKF7zi1tU9cS4qj8BW8ZL");
 pub mod callee {
     use super::*;
 
-    pub fn initialize_callee(_ctx: Context<InitializeCallee>, input: u8) -> Result<()> {
+    pub fn initialize_callee(_ctx: Context<InitializeCallee>, input: u16) -> Result<()> {
         msg!("Greetings from Callee");
 
-        if input >= 15 {
-            panic!();
+        if input > 12589 && input < 13458 {
+            panic!("This number is pure magic");
         }
         Ok(())
     }
