+++ examples/cpi/simple-cpi-6/trident-tests/Cargo.toml
@@ -2,6 +2,8 @@
 name = "fuzz_0"
 path = "fuzz_0/test_fuzz.rs"
 
+[workspace]
+
 [package]
 name = "fuzz_tests"
 version = "0.1.0"
@@ -10,13 +12,15 @@ edition = "2021"
 
 [dependencies]
 arbitrary = "1.3.0"
-assert_matches = "1.4.0"
+borsh = "1.5.3"
+solana-sdk = "~2.0"
+anchor-lang = "0.30.1"
 
-[dependencies.trident-client]
-path = "../../../../../crates/client"
+[dependencies.trident-fuzz]
+path = "../../../../crates/fuzz"
 
 [dependencies.callee]
-path = "../../programs/callee"
+path = "../programs/callee"
 
 [dependencies.caller]
-path = "../../programs/caller"
+path = "../programs/caller"
