--- examples/cpi/simple-cpi-6/trident-tests/fuzz_tests/fuzz_0/test_fuzz.rs
@@ -1,52 +1,0 @@
-use fuzz_instructions::InitializeCaller;
-use trident_client::fuzzing::*;
-mod fuzz_instructions;
-use callee::entry as entry_callee;
-use callee::ID as PROGRAM_ID_CALLEE;
-use caller::entry as entry_caller;
-use caller::ID as PROGRAM_ID_CALLER;
-use fuzz_instructions::FuzzInstruction;
-const PROGRAM_NAME_CALLEE: &str = "callee";
-const PROGRAM_NAME_CALLER: &str = "caller";
-struct InstructionsSequence;
-/// Define instruction sequences for invocation.
-/// `pre` runs at the start, `middle` in the middle, and `post` at the end.
-/// For example, to call `InitializeFn`, `UpdateFn` and then `WithdrawFn` during
-/// each fuzzing iteration:
-/// ```
-/// impl FuzzDataBuilder<FuzzInstruction> for InstructionsSequence {
-///     pre_sequence!(InitializeFn,UpdateFn);
-///     middle_sequence!(WithdrawFn);
-///}
-/// ```
-/// For more details, see: https://ackee.xyz/trident/docs/dev/features/instructions-sequences/#instructions-sequences
-impl FuzzDataBuilder<FuzzInstruction> for InstructionsSequence {
-    pre_sequence!(InitializeCaller);
-    middle_sequence!();
-    post_sequence!();
-}
-/// `fn fuzz_iteration` runs during every fuzzing iteration.
-/// Modification is not required.
-fn fuzz_iteration<T: FuzzTestExecutor<U> + std::fmt::Display, U>(
-    fuzz_data: FuzzData<T, U>,
-    config: &Config,
-) {
-    let fuzzing_program_callee = FuzzingProgram::new(
-        PROGRAM_NAME_CALLEE,
-        &PROGRAM_ID_CALLEE,
-        processor!(convert_entry!(entry_callee)),
-    );
-    let fuzzing_program_caller = FuzzingProgram::new(
-        PROGRAM_NAME_CALLER,
-        &PROGRAM_ID_CALLER,
-        processor!(convert_entry!(entry_caller)),
-    );
-    let mut client =
-        ProgramTestClientBlocking::new(&[fuzzing_program_callee, fuzzing_program_caller], config)
-            .unwrap();
-    let _ = fuzz_data.run_with_runtime(&mut client, config);
-}
-fn main() {
-    let config = Config::new();
-    fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : InstructionsSequence | { fuzz_iteration (fuzz_data , & config) ; });
-}
